services/export-service/requirements.txt
flask
psycopg2-binary==2.9.9
openpyxl
flask-cors
